+++ themes/uhsg_theme/js/degree_programme_switcher.js
@@ -3,19 +3,24 @@
   Drupal.behaviors.degreeProgrammeSwitcher = {
     attach: function(context, settings) {
       var degreeProgrammeSwitcher = '.degree-programme-switcher',
+          container = $('.degree-programme-switcher'),
           header = $('.degree-programme-switcher__header', degreeProgrammeSwitcher),
-          container = $('.degree-programme-switcher__container', degreeProgrammeSwitcher),
           toggle = $('.degree-programme-switcher__toggle', degreeProgrammeSwitcher),
           searchInput = $('input[name="name"]', degreeProgrammeSwitcher),
           toggleClass = 'collapsed',
           toggleTextClosed = Drupal.t('Change'),
-          toggleTextOpen = Drupal.t('Close');
+          toggleTextOpen = Drupal.t('Close'),
+          breakpoints = settings.breakpoints;
 
       // toggle collapsed when clicking header
       header.once().on('click', function() {
         container.toggleClass(toggleClass);
+        $('body').toggleClass('no-scroll-mobile');
         toggle.text(container.hasClass(toggleClass) ? toggleTextOpen : toggleTextClosed);
-        searchInput.focus();
+
+        if (window.matchMedia(breakpoints['small']).matches) {
+          searchInput.focus();
+        }
       });
 
       // close when clicking outside
@@ -23,6 +28,7 @@
         var clickedOutside = $(e.target).parents(degreeProgrammeSwitcher).length == 0;
         if (container.hasClass(toggleClass) && clickedOutside) {
           container.removeClass(toggleClass);
+          $('body').removeClass('no-scroll-mobile');
           toggle.text(container.hasClass(toggleClass) ? toggleTextOpen : toggleTextClosed);
         }
       })